+++ pkg/meta/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.3.1
 
-* Introduce `.unawaited` extension on `Future` used to indicates to tools that
+* Introduce `.unawaited` extension on `Future` used to indicate to tools that
   the future is intentionally not `await`-ed.
 
 ## 1.3.0